[DEV] organize some annotations
