+++ src/org/kar/archidata/annotation/apiGenerator/ApiAsyncType.java
@@ -1,4 +1,4 @@
-package org.kar.archidata.annotation;
+package org.kar.archidata.annotation.apiGenerator;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -8,7 +8,7 @@ import java.lang.annotation.Target;
 /** In case of the update parameter with String input to detect null element. */
 @Target({ ElementType.PARAMETER, ElementType.METHOD })
 @Retention(RetentionPolicy.RUNTIME)
-public @interface AsyncType {
+public @interface ApiAsyncType {
 	// Possible class values.
 	Class<?>[] value();
