+++ src/org/kar/archidata/model/UUIDGenericDataSoftDelete.java
@@ -1,9 +1,9 @@
 package org.kar.archidata.model;
 
-import org.kar.archidata.annotation.ApiGenerationMode;
 import org.kar.archidata.annotation.DataDeleted;
 import org.kar.archidata.annotation.DataNotRead;
-import org.kar.archidata.annotation.checker.ApiAccessLimitation;
+import org.kar.archidata.annotation.apiGenerator.ApiAccessLimitation;
+import org.kar.archidata.annotation.apiGenerator.ApiGenerationMode;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.annotation.Nullable;
