+++ src/org/kar/archidata/tools/AnnotationCreator.java
@@ -5,7 +5,7 @@ import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
 import java.lang.reflect.Proxy;
 
-import org.kar.archidata.annotation.ApiGenerationMode;
+import org.kar.archidata.annotation.apiGenerator.ApiGenerationMode;
 
 public class AnnotationCreator {
